Reset the `EF_VOTED` flag when re-casting a vote.
It has been observed that sometimes, the `EF_VOTED` flag isn't propagated to the client. I was unable to reproduce it, but I assume this is some kind of a RC happening when shooting, crouching or loosing helmet while casting a vote. The `EF_VOTED` seems to be persisted on the server side. This patch unsets the flag and sets it again in the subsequent frame, which I hope enforces the propagation (if this is really the issue).

This should be reverted if the bug is observed again, as it really just complicates the code in a non-obvious way.